refactor: 상품 수정 경로 변경
