+++ app/(main)/store/edit-item/[id]/page.tsx
@@ -1,4 +1,13 @@
-export default function Page({ params }: { params: { slug: string } }) {
-  console.log(params);
-  return <h1>item-edit</h1>;
-}
+import ItemEdit from './_components/ItemEdit';
+
+const ItemDetailEdit = ({ params }: { params: { id: string } }) => {
+  const { id: itemId } = params;
+
+  return (
+    <div className="flex flex-col gap-20">
+      <ItemEdit itemId={itemId} />
+    </div>
+  );
+};
+
+export default ItemDetailEdit;
