+++ app/_constants/path.ts
@@ -15,14 +15,13 @@ const pageRoute = {
     storeDetail: (id: string) => `/customer/store-detail/${id}`,
   },
   store: {
-    editItem: (id: string) => `store/edit-item/${id}`,
+    editItem: (id: string) => `/store/edit-item/${id}`,
     home: () => `/store/home`,
     login: () => `/store/login`,
     myPage: (id: string) => `/store/my-page/${id}`,
     orderDetail: (id: string) => `/store/order-detail/${id}`,
     orderList: (id: string) => `/store/order-list/${id}`,
     itemDetail: (id: string) => `/store/item-detail/${id}`,
-    itemDetailEdit: (id: string) => `/store/item-detail/${id}/edit`,
     registerItem: () => '/store/register-item',
     signUp: () => '/store/sign-up',
     storeManage: (id: string) => `/store/store-manage/${id}`,
